+++ sdk/storage/storage-queue/src/Aborter.ts
@@ -59,14 +59,14 @@ export class Aborter implements AbortSignalLike {
    *
    * @memberof Aborter
    */
-  public onabort?: (ev?: Event) => any;
+  public onabort?: (ev?: any) => any;
 
   // tslint:disable-next-line:variable-name
   private _aborted: boolean = false;
   private timer?: any;
   private readonly parent?: Aborter;
   private readonly children: Aborter[] = []; // When child object calls dispose(), remove child from here
-  private readonly abortEventListeners: Array<(this: AbortSignalLike, ev?: any) => any> = [];
+  private readonly abortEventListeners: ((this: AbortSignalLike, ev?: any) => any)[] = [];
   // Pipeline proxies need to use "abortSignal as Aborter" in order to access non AbortSignalLike methods
   // immutable primitive types
   private readonly key?: string;